# swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agrirouter Middleware",
    "description": "Middleware to access the agrirouter using a 'smaller' API. The middleware will handle the connection for all endpoints and fetches / stores messages from the agrirouter within a custom database.",
    "contact": {
      "name": "Agrirouter Middleware",
      "url": "https://github.com/agrirouter-middleware"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://github.com/agrirouter-middleware/LICENSE"
    },
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/secured/apitelemetry-platform/{endpointId}/virtual": {
      "post": {
        "tags": [
          "telemetry platform management"
        ],
        "description": "Onboard a virtual endpoint for the given telemetry platform.",
        "operationId": "telemetry-platform.onboard.virtual",
        "parameters": [
          {
            "name": "endpointId",
            "in": "path",
            "description": "The ID of the existing endpoint.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardVirtualEndpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The response if the endpoint was created successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      },
      "delete": {
        "tags": [
          "telemetry platform management"
        ],
        "description": "Revoke a virtual endpoint for the given telemetry platform.",
        "operationId": "telemetry-platform.revoke.virtual",
        "parameters": [
          {
            "name": "endpointId",
            "in": "path",
            "description": "The ID of the existing endpoint.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeVirtualEndpointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The response if the endpoint was revoked successfully.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
...
    "/secured/api/endpoint/status": {
      "post": {
        "tags": [
          "endpoint management"
        ],
        "description": "Fetch the status of an existing endpoint.",
        "operationId": "endpoint.status",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The status information for this endpoint.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
...
    "/unsecured/api/farming-software/{applicationId}/{externalEndpointId}": {
      "get": {
        "tags": [
          "farming software management"
        ],
        "description": "Create an authorization URL for the telemetry platform.",
        "operationId": "onboard.farming-software.auth-url",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "description": "The ID of the application.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalEndpointId",
            "in": "path",
            "description": "The external endpoint ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "The URL (as redirect) to authorize the creation of an endpoint within the account of the user.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/secured/api/application/{applicationId}/endpoints": {
      "get": {
        "tags": [
          "application management"
        ],
        "description": "Fetch all endpoints for an application.",
        "operationId": "application.endpoints",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "description": "The ID of the application.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The endpoints for the application.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindEndpointsForApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "The timestamp (UTC) of the response.",
            "format": "date-time"
          },
          "key": {
            "type": "string",
            "description": "The key of the error."
          },
          "errorMessage": {
            "type": "string",
            "description": "The message of the error."
          }
        },
        "description": "Generic error response to map the internal errors and provide clean API description."
      },
...
      "OnboardVirtualEndpointRequest": {
        "required": [
          "endpointName",
          "externalVirtualEndpointId"
        ],
        "type": "object",
        "properties": {
          "externalVirtualEndpointId": {
            "type": "string",
            "description": "The ID of the new virtual endpoint within the middleware."
          },
          "endpointName": {
            "type": "string",
            "description": "The list of names for the virtual endpoints."
          }
        },
        "description": "Necessary information to create the virtual endpoint."
      },
...
      "EndpointStatusRequest": {
        "required": [
          "externalEndpointIds"
        ],
        "type": "object",
        "properties": {
          "externalEndpointIds": {
            "type": "array",
            "description": "The list of external endpoint IDs.",
            "items": {
              "type": "string",
              "description": "The list of external endpoint IDs."
            }
          }
        },
        "description": "The to search for one or multiple endpoints."
      },
...
      "EndpointDto": {
        "type": "object",
        "properties": {
          "agrirouterEndpointId": {
            "type": "string",
            "description": "The ID of an endpoint."
          },
          "externalEndpointId": {
            "type": "string",
            "description": "The external ID of an endpoint."
          },
          "endpointType": {
            "type": "string",
            "description": "The type of the endpoint.",
            "enum": [
              "NON_VIRTUAL",
              "VIRTUAL"
            ]
          },
          "deactivated": {
            "type": "boolean",
            "description": "Marks an endpoint as deactivated."
          },
          "agrirouterAccountId": {
            "type": "string",
            "description": "The account ID for this endpoint."
          }
        },
        "description": "Representation of an endpoint."
      },
      "EndpointWithChildrenDto": {
        "type": "object",
        "properties": {
          "agrirouterEndpointId": {
            "type": "string",
            "description": "The ID of an endpoint."
          },
          "externalEndpointId": {
            "type": "string",
            "description": "The external ID of an endpoint."
          },
          "endpointType": {
            "type": "string",
            "description": "The type of the endpoint.",
            "enum": [
              "NON_VIRTUAL",
              "VIRTUAL"
            ]
          },
          "connectedVirtualEndpoints": {
            "type": "array",
            "description": "All the children of the endpoint.",
            "items": {
              "$ref": "#/components/schemas/EndpointDto"
            }
          },
          "deactivated": {
            "type": "boolean",
            "description": "Marks an endpoint as deactivated."
          },
          "agrirouterAccountId": {
            "type": "string",
            "description": "The account ID for this endpoint."
          }
        },
        "description": "Representation of an endpoint with all children."
      },
      "FindEndpointsForApplicationResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "The timestamp (UTC) of the response.",
            "format": "date-time"
          },
          "endpoints": {
            "type": "array",
            "description": "The endpoints for this application.",
            "items": {
              "$ref": "#/components/schemas/EndpointWithChildrenDto"
            }
          }
        },
        "description": "The response when searching for endpoints belonging to the application."
      },
...
      "RevokeVirtualEndpointRequest": {
        "required": [
          "externalEndpointIds"
        ],
        "type": "object",
        "properties": {
          "externalEndpointIds": {
            "type": "array",
            "description": "The list of endpoint IDs for the virtual endpoints.",
            "items": {
              "type": "string",
              "description": "The list of endpoint IDs for the virtual endpoints."
            }
          }
        },
        "description": "The necessary information to revoke the virtual endpoint."
      }
    },
    "securitySchemes": {
      "secured": {
        "type": "http",
        "description": "Default security scheme, used for the private flows.",
        "scheme": "basic"
      }
    }
  }
}